gvpr validTVT: take a long long instead of int
This function is always called with long long values, so lets match it with its
callers. Squashes 2 -Wconversion warnings.

+++ lib/gvpr/gprstate.c
@@ -20,8 +20,7 @@
 
 static int name_used;
 
-int validTVT(int c)
-{
+int validTVT(long long c) {
     return ((TV_flat <= c) && (c <= TV_prepostrev));
 }
